# Fix number parsing issues