--- a/src/Esprima/ArrayList.cs
+++ b/src/Esprima/ArrayList.cs
@@ -370,13 +370,18 @@ internal struct ArrayList<T> : IReadOnlyList<T>
     /// </remarks>
     public Span<T> AsSpan() => new Span<T>(_items, 0, _count);
 
+    /// <remarks>
+    /// Items should not be added or removed from the <see cref="ArrayList{T}"/> while the returned <see cref="ReadOnlySpan{T}"/> is in use!
+    /// </remarks>
+    public ReadOnlySpan<T> AsReadOnlySpan() => new ReadOnlySpan<T>(_items, 0, _count);
+
 #if NETSTANDARD2_1_OR_GREATER
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
     public T[] ToArray()
     {
 #if NETSTANDARD2_1_OR_GREATER
-        return AsSpan().ToArray();
+        return AsReadOnlySpan().ToArray();
 #else
         if (_count == 0)
         {