… test/Esprima.Tests/ScannerTests.cs