Fixes several bugs in number parsing (especially bigints):
* `0o` bigint literals should be parsed as octal numbers (were incorrectly parsed as decimal).
* `0b` and `0o` literals which are larger than a 64-bit integer are valid and should be parsed.
* `0x<256 or more zeros>` should not result in `double.PositiveInfinity` (see https://github.com/sebastienros/esprima-dotnet/compare/fix/number-parsing?expand=1#diff-5c649e707747c206cd95c805966b187aa0e3f0725f25f51aef684f8254ad0149L1065).
* Legacy octal numbers with group separators (e.g. `0777_777`) should be a syntax error (`FormatException` was thrown).
* Fractions with omitted leading zero and trailing 'n'  (e.g. `.17n`) should be a syntax error (no exception was thrown).

Also fixes a few minor bugs (e.g. in `AstToJsonConverter` when numeric literal value is `Infinity`).

&plus; Some micro-optimization.