@adams85 2023-08-18
## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |ParseProgram|angular-1.2.5|8,737.0 μs|122.22 μs|4095.44 KB|
| **New** |	|	| **8,604.0 μs (-2%)** | **781.46 μs** | **4095.47 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|1,099.7 μs|24.22 μs|647.78 KB|
| **New** |	|	| **1,113.5 μs (+1%)** | **64.11 μs** | **647.78 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|6,561.9 μs|184.62 μs|3540.13 KB|
| **New** |	|	| **6,597.9 μs (+1%)** | **292.46 μs** | **3540.13 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|10,105.1 μs|780.14 μs|5609.22 KB|
| **New** |	|	| **10,184.9 μs (+1%)** | **535.63 μs** | **5609.24 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|5,262.7 μs|239.04 μs|2938.38 KB|
| **New** |	|	| **5,195.0 μs (-1%)** | **76.77 μs** | **2938.37 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|953.8 μs|24.72 μs|555.5 KB|
| **New** |	|	| **927.7 μs (-3%)** | **169.11 μs** | **555.5 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|4,521.4 μs|112.76 μs|2714.65 KB|
| **New** |	|	| **4,486.3 μs (-1%)** | **226.24 μs** | **2714.65 KB (0%)** |