@lahma 2023-08-18
@adams85  a bit unrelated, but what are your thoughts about 3.0 maturity, should the big release be made and leave smaller bug fixes to maintenance releases. Or do you see some API work ahead still?