@adams85 2023-08-18
Nope, as far as I'm concerned, no new features are planned. IMO the parser is pretty solid now, I only know of one area (syntax error detection and reporting) which would need some love but that would take a lot of work for not pretty little gain. I can't see that work be done soon (at least, in the current codebase - but more on this later.) So, I don't think that minor bugs and refinements should block 3.0.

There are a few cleanup tasks in my todo list like [this](https://github.com/sebastienros/esprima-dotnet/issues/303). Hopefully, I can take care of them in the next week(s). I'll add those tasks to the [3.0 milestone](https://github.com/sebastienros/esprima-dotnet/milestone/1) so we can keep track of the progress. I suggest you do the same if you have anything in mind that you'd like to make its way into 3.0